# src/main/java/com/example/bankcards/dto/CardResponse.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.CardStatus;

public record CardResponse(
        Long id,
        Long userId,
        String number,
        String cardholderName,
        Short expirationMonth,
        Short expirationYear,
        CardStatus status
) {
}